+++ src/plugins/plot/stackedPlot/stackedPlotConfigurationInterceptor.js
